+++ rspec-core/lib/rspec/core/option_parser.rb
@@ -178,11 +178,7 @@ module RSpec::Core
         end
 
         parser.on('-w', '--warnings', 'Enable ruby warnings') do
-          if Object.const_defined?(:Warning) && Warning.respond_to?(:[]=)
-            # :nocov: on older Ruby without Warning
-            Warning[:deprecated] = true
-            # :nocov:
-          end
+          Warning[:deprecated] = true
           $VERBOSE = true
         end